Wound Counter
This program takes in from the user the number dice being rolled, base number for hits to hit, base number for wounds to wound, and number for saving throws. Following the rules laid out in the Warhammer 40k rule book the program determines which of the dice pass all conditions to be consider a damage to defend unit.